Drogue separation point changed
changed nose cone mass to 0 grams to reduce stability. can be increased in the future if need be.

recovery hardware had an equal diameter to the OD of the tube.

added in threaded rod

increased length of nosecone coupler to 12", moved cloverbulk head to halfway point of coupler.

reduced the length of the avionics bay coupler to 32", 4" overlap with drogue bay and 6" overlap at separation point with booster section

right now this version is under-stable- to be fixed soon